Time: 324 ms (44.66%), Space: 136.2 MB (44.66%) - LeetHub

+++ 2285-maximum-total-importance-of-roads/2285-maximum-total-importance-of-roads.cpp
@@ -16,7 +16,7 @@ public:
         
         for(int i=0; i<n; i++) {
             if(arr[i]) {
-                ans += (long long)(n-i) * (long long)arr[i];
+                ans += (long long)(n-i) * arr[i];
             }
         }
